Backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gTTS
sarvamai
google-generativeai